requirements.txt
requests==2.32.4
pycryptodome==3.23.0
tqdm==4.67.1